Checker/requirements.txt
pymysql
sqlalchemy
python-dotenv
requests
PyMuPDF
Pillow
pyzbar
pdf2image